Moving main AGauge control into new namespace
AGauge control class is now in AGauge.Controls namespace.
Aiming for organization and modularization here. Also planning to create some useful composite controls.
