+++ AGauge/Composites/README.md
@@ -1,0 +1,4 @@
+﻿# Purpose
+Composites are collections of controls, for the purose of being used as a single control to the end-user.
+
+These are available through the `AGauge.Composites` namespace.
